feat: Add Chrome and Firefox manifest files and update build scripts for extension
