+++ manifest.chrome.json
