+++ package.json
@@ -3,8 +3,12 @@
   "version": "0.1.0",
   "description": "Auto Translation Extension",
   "scripts": {
-    "build": "webpack --mode production",
-    "watch": "webpack --watch --mode development",
+    "build:chrome": "webpack --config webpack.chrome.js --mode production",
+    "build:firefox": "webpack --config webpack.firefox.js --mode production",
+    "build:all": "npm run build:chrome && npm run build:firefox",
+    "watch": "webpack --config webpack.chrome.js --watch --mode development",
+    "watch:chrome": "webpack --config webpack.chrome.js --watch --mode development",
+    "watch:firefox": "webpack --config webpack.firefox.js --watch --mode development",
     "lint": "eslint src/**/*.js",
     "format": "prettier --write src/**/*.js"
   },
